requirements.txt
azure-storage==0.33.0
azure-servicebus==0.20.2
